Improve config
