+++ src/discord/commands/PineconeModal.ts
@@ -1,6 +1,7 @@
-import {getConfig, trySavingPineconeOptions} from "../../core/config";
+import {getConfig} from "../../core/config";
 import {EmbedBuilder, TextInputStyle} from "discord.js";
 import {defineModal} from "./DefineModal";
+import {trySavingPineconeOptions} from "../../core/pinecone";
 
 const pineconModalId = 'SET-PINECONE-MODAL';
