+++ src/discord/listeners/ready.ts
@@ -5,6 +5,9 @@ import {InitializeThreads} from "./ready/initializeThreads";
 import {logMessage} from "../../utils/logMessage";
 import {db} from "../../database/db";
 import {getEnv} from "../../utils/GetEnv";
+import {getConfig} from "../../core/config";
+import {mainServerId} from "../../core/MainServerId";
+import {getOpenAIForId} from "../../core/GetOpenAIForId";
 
 const MAIN_SERVER_ID = getEnv('MAIN_SERVER_ID');
 
@@ -30,6 +33,25 @@ export default (client: Client): void => {
         const guilds = await client.guilds.fetch();
 
         logMessage(`Currently in ${guilds.size} guild${guilds.size !== 1 ? 's' : ''}.`);
+
+        await getConfig();
+
+        let openaiSuccess = false;
+
+        const openai = await getOpenAIForId(mainServerId);
+        if (openai) {
+            try {
+                const models = await openai.listModels();
+
+                if (models) {
+                    openaiSuccess = true;
+                }
+            } catch (e) {
+                logMessage('Cannot use openai:', e);
+            }
+        }
+
+        logMessage(`Openai connection: ${openaiSuccess ? 'good' : 'bad'}!`);
 
         StartListeningToMessages(client);
         InitializeThreads();
